re-add classes for noclasses
I use a custom class for each style. The most common
need for css seems to be night mode. Usually I want a really
dark background but such a background doesn't work for
all themes, e.g. keywords in the borland theme are dark
blue and hardly readable on a dark background